requirements.txt
google-generativeai
python-dotenv
newspaper3k
nltk
openai